Breaking up the plethora of methods
Breaking up the plethora of methods into more easily digested
and findable files.
