+++ curator/optimize.py
@@ -1,0 +1,23 @@
+def optimize_index(client, index_name, max_num_segments=2, **kwargs):
+    """
+    Optimize (Lucene forceMerge) index to ``max_num_segments`` per shard
+
+    :arg client: The Elasticsearch client connection
+    :arg index_name: The index name
+    :arg max_num_segments: Merge to this number of segments per shard.
+    """
+    if csv_check(index_name):
+        logger.error("Must specify only a single index as an argument.")
+        return True
+    if index_closed(client, index_name): # Don't try to optimize a closed index
+        logger.info('Skipping index {0}: Already closed.'.format(index_name))
+        return True
+    else:
+        shards, segmentcount = get_segmentcount(client, index_name)
+        logger.debug('Index {0} has {1} shards and {2} segments total.'.format(index_name, shards, segmentcount))
+        if segmentcount > (shards * max_num_segments):
+            logger.info('Optimizing index {0} to {1} segments per shard.  Please wait...'.format(index_name, max_num_segments))
+            client.indices.optimize(index=index_name, max_num_segments=max_num_segments)
+        else:
+            logger.info('Skipping index {0}: Already optimized.'.format(index_name))
+            return True
